:sparkles: 使得forecast_list输出文本经过格式化
