+++ __main__.py
@@ -45,7 +45,7 @@ if __name__ == "__main__" and len(sys.argv) < 2:
             print("\nHoliday Predictor / 假期预测器 已结束运行。感谢你的使用。\n")
             break
 
-        if input_list[0] in ("forecast_list", "fcls"): # 处理用户发起的 forecast_list 指令请求。
+        if input_list[0] in ("forecast_list", "fcls"):  # 处理用户发起的 forecast_list 指令请求。
             print(calculation_forecast_list(input_list))
             execution_flag = True
